Fix windows roles reading from puppet
Puppet agent needs to read the role and set it on Windows servers before
being able to provision the node. In order to know the role to which the
server is assigned it needs to contact puppetmaster-gui server in the
roles route. The previous implementation just searched for the hostname
directly in the database. Sometimes for windows server if the domain is
different puppet may think that the hostname of the machine is not the
same as the one written in the hiera database trough puppetmaster-gui.
Consider the example:
 - Resolvable hostname: public-apps.example.com
 - Win server hostname: public-apps
 - Hostname wit domain: public-apps.example.local
All these three examples contain one common part and that is the string
that represents the machine name "public-apps". The fix just strips
everything after the dot in the hostname and searches by the first part
of the string. Since the server names must be unique then this will work
in any mentioned case.

Resolves PROD-1927.