Change puppet msi download link
During bootstrap process for Windows servers, puppetmaster-gui was not
able to download puppet MSI package for windows since the download link
has been changed. This resulted in completeley broken windows
provisioning that could not even start.

By chnging the link to the latest puppet 3.8.7 we should be good to go
at least up to Windows server 2016.

Resolves PROD-1928.